Improve setup.py, cleanup requirement and remove setup.cfg file.
`setup_requires` needs to be specified in *both* `setup.py` and
`pyproject.toml`.

The `requirements.txt` file should use `-e .` to install the module (and
thus get the requirements). If a local or non-PyPi version of a module
should be used to satisfy a requirement for either `setup_requires` or
`install_requires` then it should also be specified in the
`requirements.txt` file.

Using `setup.py` is much clearer than `setup.cfg` and it looks like the
file is going to get replaced by `pyproject.toml` in the future, see
https://github.com/pypa/setuptools/issues/1688

The generated Cython C file is included in the source distribution, so
only people building from the git repository need to have Cython
installed (fixing import issue).

Signed-off-by: Tim 'mithro' Ansell <me@mith.ro>