gha: Only allow one jobs at once.
The wheels seem to be randomly failing on GitHub Actions and I'm
randomly guessing it is due to an out of memory error. See
https://github.com/SymbiFlow/fasm/issues/77

Don't use parallel build if `CMAKE_BUILD_PARALLEL_LEVEL` is in the
environment and set `CMAKE_BUILD_PARALLEL_LEVEL=1` in the GitHub Action
environment.

Hopefully fixes https://github.com/SymbiFlow/fasm/issues/77

Signed-off-by: Tim 'mithro' Ansell <tansell@google.com>